+++ lib/get-applicable-versions.js
@@ -1,3 +1,4 @@
+import { fileURLToPath } from 'url'
 import path from 'path'
 import { reduce, sortBy } from 'lodash-es'
 import { allVersions } from './all-versions.js'
@@ -5,12 +6,10 @@ import versionSatisfiesRange from './version-satisfies-range.js'
 import checkIfNextVersionOnly from './check-if-next-version-only.js'
 import dataDirectory from './data-directory.js'
 import encodeBracketedParentheses from './encode-bracketed-parentheses.js'
-
-const featuresDir = path.join('data', 'features')
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
+const featuresDir = path.posix.join(__dirname, '../data/features')
 
 let featureData = null
-
-const allVersionKeys = Object.keys(allVersions)
 
 // return an array of versions that an article's product versions encompasses
 function getApplicableVersions(frontmatterVersions, filepath) {
@@ -20,7 +19,7 @@ function getApplicableVersions(frontmatterVersions, filepath) {
 
   // all versions are applicable!
   if (frontmatterVersions === '*') {
-    return allVersionKeys
+    return Object.keys(allVersions)
   }
 
   if (!featureData) {
@@ -80,7 +79,7 @@ function getApplicableVersions(frontmatterVersions, filepath) {
 
   // Sort them by the order in lib/all-versions.
   const sortedVersions = sortBy(applicableVersions, (v) => {
-    return allVersionKeys.indexOf(v)
+    return Object.keys(allVersions).indexOf(v)
   })
 
   return sortedVersions
